Store & access instrumentation as series of events.
It was a series of events before but was accessed as a big blob of bytes, which is much fun to use.
Also: turns out the strings returned from ICorProfilerInfo include the nul terminator in the
length, which caused some confusion.

+++ ReactivityProfiler/RxMethodInstrumentation.cpp
@@ -286,7 +286,7 @@ void MethodBodyInstrumenter::InstrumentCall(ObservableCallInfo& call, CMetadataE
         ATLTRACE("Got locals sig: %s", FormatBytes(localsSigBlob).c_str());
         LocalsSignatureReader localsSigReader(localsSigBlob);
         int existingLocalsCount = localsSigReader.GetCount();
-        int argCount = call.m_argIsObservable.size(); // not necessarily all args, but the ones we're dealing with
+        int argCount = static_cast<int>(call.m_argIsObservable.size()); // not necessarily all args, but the ones we're dealing with
         std::vector<SignatureBlob> argTypeSpans;
         std::transform(
             call.m_argTypeSpan.begin(), call.m_argTypeSpan.end(),
